==> 4-1-preprocessing_indisponibilites_proc.ipynb <==
# Traitement du fichier des indisponibilités des installations de production d'électricité en France fourni par RTE pour
# calcul des capacités de production disponible par date

"""
Created on Mon May 27 12:42:29 2024

@author: Thierry ALLEM
"""

import numpy as np
import pandas as pd
import datetime

#==========================  CHARGEMENT des fichiers d'indisponibilités des moyens de production de 2013 à 2023 ===============================================================================
ind_prod_2013 = pd.read_csv('DonneesIndisponibilitesProduction_2013.csv', sep=';',encoding='latin-1')
ind_prod_2014 = pd.read_csv('DonneesIndisponibilitesProduction_2014.csv', sep=';',encoding='latin-1')
ind_prod_2015 = pd.read_csv('DonneesIndisponibilitesProduction_2015.csv', sep=';',encoding='latin-1')
ind_prod_2016 = pd.read_csv('DonneesIndisponibilitesProduction_2016.csv', sep=';',encoding='latin-1')
ind_prod_2017 = pd.read_csv('DonneesIndisponibilitesProduction_2017.csv', sep=';',encoding='latin-1')
ind_prod_2018 = pd.read_csv('DonneesIndisponibilitesProduction_2018.csv', sep=';',encoding='latin-1')
ind_prod_2019 = pd.read_csv('DonneesIndisponibilitesProduction_2019.csv', sep=';',encoding='latin-1')
ind_prod_2020 = pd.read_csv('DonneesIndisponibilitesProduction_2020.csv', sep=';',encoding='latin-1')
ind_prod_2021 = pd.read_csv('DonneesIndisponibilitesProduction_2021.csv', sep=';',encoding='latin-1')
ind_prod_2022 = pd.read_csv('DonneesIndisponibilitesProduction_2022.csv', sep=';',encoding='latin-1')

# ===========================INFORMATIONS sur les dataframes==================================================================================================================================
print("2013")
ind_prod_2013.info()
print("2014")
ind_prod_2014.info()
print("2015")
ind_prod_2015.info()
print("2016")
ind_prod_2016.info()
print("2017")
ind_prod_2017.info()
print("2018")
ind_prod_2018.info()
print("2019")
ind_prod_2019.info()
print("2020")
ind_prod_2020.info()
print("2021")
ind_prod_2021.info()
print("2022")
ind_prod_2022.info()

# Export des informations vers un fichier Excel

# Création d'un dictionnaire des dataframes
dataframes = {
    'ind_prod_2013': ind_prod_2013, 'ind_prod_2014': ind_prod_2014,'ind_prod_2015': ind_prod_2015,'ind_prod_2016': ind_prod_2016,
    'ind_prod_2017': ind_prod_2017,'ind_prod_2018': ind_prod_2018, 'ind_prod_2019': ind_prod_2019, 'ind_prod_2020': ind_prod_2020,
    'ind_prod_2021': ind_prod_2021, 'ind_prod_2022': ind_prod_2022 }

# Initialisation d'un DataFrame vide pour stocker les informations
df_info = pd.DataFrame()

# Parcours de chaque dataframe pour récupérer les informations
for df_name, df in dataframes.items():
    # Obtention du nombre de données non-null par colonne
    non_null_counts = df.count()
    # OObtention des types de données (DTypes) par colonne
    dtypes = df.dtypes
    # Création d'un DataFrame temporaire avec les informations
    df_temp = pd.DataFrame({
        'Non-Null Counts': non_null_counts,
        'DTypes': dtypes
    })
    # Renommage des colonnes du DataFrame temporaire avec le nom du dataframe
    df_temp.columns = [f'{df_name} - Non-Null Counts', f'{df_name} - DTypes']
    # Ajout du DataFrame temporaire au DataFrame global
    df_info = pd.concat([df_info, df_temp], axis=1)

# Écriture du DataFrame global dans un fichier Excel
df_info.to_excel('rapport_dataframes_indisponibilites_13_22.xlsx', index=False)

# ===========================SUPPRESSION DES COLONNES superflues de ind_prod_2021 et ind_prod_2023 =====================================================================================
colonnes_to_drop2021 = [19]
ind_prod_2021.drop(columns = ind_prod_2021.columns[colonnes_to_drop2021], inplace =True)

# =========================== CONCATENATION de toutes les données d'indisponibilités de production de 2013 à 2023 ======================
ind_prod=pd.concat([ind_prod_2013, ind_prod_2014, ind_prod_2015, ind_prod_2016, ind_prod_2017, ind_prod_2018, ind_prod_2019, ind_prod_2020, ind_prod_2022], axis = 0) 
ind_prod.head()

# =========================== RENOMMAGE DES COLONNES ==================================================================================================================================

    # Dictionnaire avec les nouveaux noms des colonnes

dictionnaire_entetes = {'Identifiant du message':'id_message',
                'Date et heure de la publication':'date_heure_publication',
                 'Date de début' : 'date_debut',                
                'Date de fin prévisionnelle' : 'date_fin_previsionnelle',
                "Type d'indisponibilité": 'type_indisponibilite',
                'Filière de production' : 'filiere_production',
                "Type d'évènement" : 'type_evenement',
                'Nom du groupe ou de la centrale de production' :'nom_centrale',
                'Code EIC du groupe ou de la centrale de production' : 'code_eic_centrale',
                'Société' : 'societe',
                'Code EIC de la société' : 'code_eic_societe',
                'Capacité installée' : 'capacite_installee',
                'Capacité disponible' : 'capacite_disponible',
                'Capacité indisponible' : 'capacite_indisponible',
                "Cause de l'arrêt" : 'cause_arret',
                'Remarques' : 'remarques',
                'Statut' : 'statut',
                'Unité de mesure' : 'unite_de_mesure',
                'Code localisation' : 'code_localisation'}
                
ind_prod=ind_prod.rename(dictionnaire_entetes, axis = 1)

# ========================== REMPLACEMENT DES NaN de la colonne 'code_eic_centrale' par 'NR' (non renseigné) ===============================================================================
ind_prod['code_eic_centrale'] = ind_prod['code_eic_centrale'].fillna('NR')

# =========================== RAPPROCHEMENTS DES CODES EIC du dataframe des indisponibilités avec le dataframe du registre des centrales ===================================================
# de production afin d'ajouter une colonne 'région_ncc' aux données d'indisponibilités

    # Récupération du fichier des capacités de production pré-processé nommé 'reg_prod_cleaned'
reg_prod_cleaned = pd.read_csv('registre_prod_cleaned.csv', sep=';',encoding='latin-1')

    # Récupération des noms des régions dans le dataframe des indisponibilités et remplacement
    # des valeurs manquantes par 'NT' (non trouvée)
        # Création d'un dictionnaire de correspondance
eic_to_region = dict(zip(reg_prod_cleaned['code_EIC'], reg_prod_cleaned['region']))

# AJOUT DE LA COLONNE 'region' dans ind_prod
ind_prod.insert(0, 'region', ind_prod['code_eic_centrale'].map(eic_to_region).fillna('NT'))

#  =========================== ANALYSE DES LIGNES pour lesquelles la centrale de production n'a pu être rattachée à une région (valeurs 'NT') =============================================
    # Nombre de lignes avec une valeur 'NT' dans la colonne 'region'
nb_NT = (ind_prod['region'] == 'NT').sum()
print("Nombre d'installations sans région renseignée : ", nb_NT)

    # Nombre de valeurs uniques de 'nom_centrale' pour lesquels region =NTca
        # Filtration des lignes où 'region' est 'NT'
nt_region = ind_prod[ind_prod['region'] == 'NT']

    # Comptage des valeurs uniques dans la colonne 'nom_centrale'
unique_count = nt_region['nom_centrale'].nunique()

print(f"Nombre de valeurs uniques du nom de l'installation pour lesquelles la 'region' n'est pas renseignée : {unique_count}")

    # Somme des capacités installées pour lesquelles 'région' = 'NT'
    
        # Remplacement des virgules par des points
ind_prod['capacite_installee'] = ind_prod['capacite_installee'].str.replace(',', '.')
    # Conversion de la colonne 'Capacité installée' en float
ind_prod['capacite_installee'] = ind_prod['capacite_installee'].astype(float)

capacites_nt = ind_prod.loc[ind_prod['region'] == 'NT', 'capacite_installee'].sum()
print("La somme des valeurs de la colonne 'capacite_installee' pour 'region' = 'NT' est :", capacites_nt)
print("La somme des capacités installée des installations sans région attribuée n'est pas nulle")
print("Il convient donc de compléter ces régions manquantes. Pour cela, une recherche manuelle par code EIC")
print("sera faite via le fichier Excel issu de l'exportation de 'ind_prod' et les moteurs de recherche d'Internet")

    # Export du dataframe des indisponibilités de 2013 à 2023 dans un fichier Excel

file_export ='Indisponibilites_2013_2023_regions.xlsx'

ind_prod.to_excel(file_export)

print("Rapport préliminaire d'exploration des colonnes du dataframe 'ind_prod, exporté avec succès dans le fichier Excel 'Indisponibilites_2013_2023_regions.xlsx'")

    # Import du dataframe 'ind_prod_compl' issu du remplissage manuel des valeurs manquantes de 'region' 
ind_prod_comp = pd.read_csv('Indisponibilites_2013_2023_regions_compl.csv', sep=';',encoding='latin-1')   

print("Il reste des installations pour lesquelles le nom de la région n'a pu être attribuée")

# RECHERCHE ET TRAITEMENT DES DOUBLONS

doublons = ind_prod_comp.duplicated(keep=False)

# Filtration des doublons et copie dans un nouveau DataFrame
nouveau_dataframe_doublons = ind_prod_comp[doublons]

nombre_de_doublons = ind_prod_comp.duplicated(keep=False).sum()
print(f"Nombre total de lignes en doublons : {nombre_de_doublons}")

duplicates = ind_prod_comp[ind_prod_comp.duplicated()]
print("Lignes dupliquées :")
print(duplicates)

# Suppression des doublons (en gardant la première occurrence)
col_subset = ['date_debut', 'date_fin_previsionnelle', 'filiere_production', 'code_eic_centrale','capacite_disponible', 'capacite_indisponible']

ind_prod_comp.drop_duplicates(keep='first',subset = col_subset, inplace=True)

nombre_de_doublons2 = ind_prod_comp.duplicated(keep=False).sum()

print(f"Nombre total de lignes en doublons après suppressions: {nombre_de_doublons2}")


#  =================================== ANALYSE DES LIGNES de 'ind_prod_comp' portant la valeur 'NT' dans 'region'

nb_NT2 = (ind_prod_comp['region'] == 'NT').sum()
print("Nombre d'installations sans région renseignée : ", nb_NT2)

    # Nombre de valeurs uniques de 'nom_centrale' pour lesquels region =NT
        # Filtration des lignes où 'region' est 'NT'
nt_region2 = ind_prod_comp[ind_prod_comp['region'] == 'NT']

        # Comptage des valeurs uniques dans la colonne 'nom_centrale'
unique_count2 = nt_region2['nom_centrale'].nunique()

print(f"Nombre de valeurs uniques du nom de l'installation où 'region' a pour valeur 'NT' : {unique_count2}")

    # Somme des capacités installées pour lesquelles région = 'NT'
        # Remplacement des virgules par des points
ind_prod_comp['capacite_installee'] = ind_prod_comp['capacite_installee'].str.replace(',', '.')
    # Conversion de la colonne 'Capacité installée' en float
ind_prod_comp['capacite_installee'] = ind_prod_comp['capacite_installee'].astype(float)

capacites_nt = ind_prod_comp.loc[ind_prod_comp['region'] == 'NT', 'capacite_installee'].sum()
print("La somme des valeurs de la colonne 'capacite_installee' pour 'region' = 'NT' après est :", capacites_nt)

print("On note que la capacité installée totale des installations sans région attribuée est nulle; on peut donc supprimer les lignes correspondantes")
# Filtration des lignes où 'region' est 'NT' et suppression
ind_prod_comp.drop(ind_prod_comp[ind_prod_comp['region'] == 'NT'].index, inplace = True)

        # Filtration des lignes où 'region' est 'NT'
nt_region3= ind_prod_comp[ind_prod_comp['region'] == 'NT']

        # Comptage des valeurs uniques dans la colonne 'nom_centrale'
unique_count3 = nt_region3['nom_centrale'].nunique()

print(f"Nombre de valeurs uniques du nom de l'installation où 'region' a pour valeur 'NT' après suppressions : {unique_count3}")

print("Lignes pour lesquelles 'region' = 'NT' et 'somme des capacités installées' = 0,  supprimées")
ind_prod_comp.head()

ind_prod_comp.head()

ind_prod_comp.info()

#=============================== CONVERSION DES FORMATS DE COLONNES ==============================================================================================

    # Conversion des colonnes 'Capacités' de object en float
ind_prod_comp['capacite _installee'] = ind_prod_comp['capacite_installee'].astype('float64')
ind_prod_comp['capacite_disponible'] = ind_prod_comp['capacite_disponible'].astype('float64')
ind_prod_comp['capacite_indisponible'] = ind_prod_comp['capacite_indisponible'].str.replace(',', '.').astype('float64')

    # Conversion des colonnes 'dates'
ind_prod_comp['date_heure_publication'] = pd.to_datetime(ind_prod_comp['date_heure_publication'], format='%d/%m/%Y %H:%M').dt.strftime('%Y-%m-%d %H:%M')
ind_prod_comp['date_heure_publication'] = pd.to_datetime(ind_prod_comp['date_heure_publication'])
ind_prod_comp['date_debut'] = pd.to_datetime(ind_prod_comp['date_debut'], format='%d/%m/%Y %H:%M').dt.strftime('%Y-%m-%d %H:%M')
ind_prod_comp['date_debut'] = pd.to_datetime(ind_prod_comp['date_debut'])
ind_prod_comp['date_fin_previsionnelle'] = pd.to_datetime(ind_prod_comp['date_fin_previsionnelle'], format='%d/%m/%Y %H:%M').dt.strftime('%Y-%m-%d %H:%M')
ind_prod_comp['date_fin_previsionnelle'] = pd.to_datetime(ind_prod_comp['date_fin_previsionnelle'])

# =============================== RENOMMAGE DES VALEURS UNIQUES =================================================================================================
    # ************ NOM DES REGIONS *************************************************************************************************************************
print("Valeurs uniques de 'region' AVANT renommage :", ind_prod_comp['region'].unique())
dictionnaire_region= {"Provence-Alpes-Côte d'Azur":"PROVENCE ALPES COTE D AZUR",
                      "Bourgogne-Franche-Comté":"BOURGOGNE FRANCHE COMTE",
                      "Bretagne":"BRETAGNE",
                      "Normandie":"NORMANDIE",
                      "Nouvelle-Aquitaine":"NOUVELLE AQUITAINE",
                      "Hauts-de-France":"HAUTS DE FRANCE",
                      "Île-de-France":"ILE DE FRANCE", 
                      "Auvergne-Rhône-Alpes":"AUVERGNE RHONE ALPES",
                      "Grand Est":"GRAND EST",
                      "Centre-Val de Loire":"CENTRE VAL DE LOIRE",
                      "Occitanie":"OCCITANIE",
                      "Pays de la Loire":"PAYS DE LA LOIRE"}
ind_prod_comp['region']=ind_prod_comp['region'].replace(dictionnaire_region)
print("Valeurs uniques de 'region' APRES renommage :",ind_prod_comp['region'].unique())

# Suppression de la colonne d'index 21 de 'capacite_installee' en doublon
colonnes_to_drop = [21]
ind_prod_comp.drop(columns = ind_prod_comp.columns[colonnes_to_drop], inplace =True)

    # ************** NOM DES FILIERES DE PRODUCTION **********************************************************************************************
print("Valeurs uniques de la colonne 'filiere_production':", ind_prod_comp['filiere_production'].unique())

print("Nombre de valeurs manquantes de la colonne 'filiere_production' :",ind_prod_comp['filiere_production'].isna().sum())
ind_prod_comp.dropna(subset=['filiere_production'], inplace=True)
print("Nombre de valeurs manquantes de la colonne 'filiere_production' :",ind_prod_comp['filiere_production'].isna().sum())

             # Remplacement de la valeur unique 'Autre'
print("Après recherches, la valeur 'Autres' de la colonne 'filiere_production' est rattachée à la centrale thermique à gaz CyCoFos")
print("La valeur 'Autre' sera donc remplacée par la valeur 'thermique'")          

            #  Ajout d'une colonne portant le nom de la filière conformément à ceux renseignés dans le fichier 'eco2mix-regional-cons-def'
            # Insertion d'une nouvelle colonne 'filiere_globale' à l'index 7
ind_prod_comp.insert(7, 'filiere_globale', ind_prod_comp['filiere_production'])

        # Affectation des valeurs de 'region_ncc' à 'region_abr'
ind_prod_comp['filiere_globale'] = ind_prod_comp['filiere_production']
 # Renommage des valeurs de 'filiere_production' pour équivalence avec les données de 'eco2mix-regional-cons-def'
dictionnaire_region_abr= {'Gaz':'thermique',
                          'Nucléaire': 'nucleaire',
                          'Fioul' : 'thermique',
                          'Hydraulique STEP': 'hydraulique',
                          'Hydraulique lacs' : 'hydraulique',
                          'Charbon': 'thermique',
                          "Hydraulique fil de l'eau / éclusée" : 'hydraulique',
                          'Autre' : 'thermique',
                          'Marin' : 'hydraulique',
                          'Biomasse' : 'bioenergies',
                          'Éolien en mer' : 'eolien'}
ind_prod_comp['filiere_globale']=ind_prod_comp['filiere_globale'].replace(dictionnaire_region_abr)
print(ind_prod_comp['filiere_globale'].unique())

apercu = ind_prod_comp.head(20)

# GESTION DES DATES

    # Valeurs absentes de la colonne 'date_debut'
print("Nombre de valeurs manquantes de la colonne 'date_debut' :",ind_prod_comp['date_debut'].isna().sum())

    # Valeurs absentes de la colonne 'date_fin_previsionnelle'
print("Nombre de valeurs manquantes de la colonne 'date_fin_previsionnelle' :",ind_prod_comp['date_fin_previsionnelle'].isna().sum())

print("En l'absence d'informations suffisantes sur ces dates de fin prévisionnelles absentes, le choix est fait de faire une sauvegarde")
print(" de ces données dans un fichier séparé afin de les exploiter éventuellement en cas de détection d'anomalies lors des analyses ultérieures")
print("puis de les supprimer du dataframe ind_prod_comp ")

        # Export des lignes avec 'date_fin_previsionnelle' absentes dans un fichier Excel pour traitement
ind_prod_comp_filtered = ind_prod_comp.loc[ind_prod_comp['date_fin_previsionnelle'].isna()]

ind_prod_comp_filtered.to_excel('ind_prod_comp__date_fin_manquantes.xlsx', index=False)
# Filtration des lignes où 'date_fin_previsionnelle' est manquante et suppression
ind_prod_comp = ind_prod_comp.dropna(subset=['date_fin_previsionnelle'])
print("Nombre de valeurs manquantes de la colonne 'date_fin_previsionnelle'après suppressions:",ind_prod_comp['date_fin_previsionnelle'].isna().sum())

# CALCUL DE LA DUREE DES INDISPONIBILITES EN MINUTES
        # Insertion d'une nouvelle colonne 'duree_inter_minutes' à l'index 7
ind_prod_comp.insert(7, 'duree_inter_minutes', '0')

        # calcul des durées d'interruption en minutes
ind_prod_comp['duree_inter_minutes'] = (ind_prod_comp['date_fin_previsionnelle'] - ind_prod_comp['date_debut']).dt.total_seconds() / 60

apercu2 = ind_prod_comp.head(20)
ind_prod_comp.info()


ind_prod_comp.isna().sum()

ind_prod_comp['statut'].unique()
ind_prod_comp_copy= ind_prod_comp.copy()

# Suppression des lignes où la valeur de la colonne 'statut' est manquante

ind_prod_comp = ind_prod_comp.dropna(how='all', subset=['statut'])

# Suppression des lignes où la colonne 'statut' a la valeur 'Annulée' ou 'Inactive'
ind_prod_comp = ind_prod_comp.drop(ind_prod_comp[ind_prod_comp['statut'].isin(['Annulée', 'Inactive'])].index)

ind_prod_comp['statut'].unique()

ind_prod_comp = ind_prod_comp.replace({'': np.nan, ' ': np.nan, '      ':np.nan})

ind_prod_comp.isna().sum()

 # Export du dataframe final, après traitements, des indisponibilités de 2013 à 2023 dans un fichier csv et Excel

indisponibilites_cleaned = ind_prod_comp.copy()
file_export_csv ='Indisponibilites_2013_2023_regions_cleaned.csv'

ind_prod_comp.to_csv(file_export_csv,sep=';',encoding='latin-1')

file_export_xlsx='Indisponibilites_2013_2023_regions_cleaned.xlsx'
ind_prod_comp.to_excel(file_export_xlsx)

ind_prod_comp.info()

# Lors de l'examen du fichier CSV depuis Excel, on constate de très nombreuses lignes sur lesquelles les valeurs de 'capacité_indisponible' sont absente; pourtant elles ne sont
# pas détectées lors de l'isna. Il y a aussi des lignes de commentaires insérés entres les lignes de données. L'ensemble de ces lignes est donc supprimé manuellement via la fonction
# filtrage de Excel puis le fichier réimporté ci-dessous

 # REIMPORT DU DATAFRAME AVEC LE FICHIER TRAITE MANUELLEMMENT POUR ENLEVER LES LIGNES DE COMMENTAIRES  ET ABSENTES
ind_prod_comp = pd.read_csv('Indisponibilites_2013_2023_regions_def.csv', sep=';',encoding='latin-1', header = 1)

#            SUPPRESSION de la colonne 'Unnamed: 0' de 'ind_prod_comp' ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
colonnes_to_drop = [0]
ind_prod_comp.drop(columns = ind_prod_comp.columns[colonnes_to_drop], inplace =True)

ind_prod_comp_copy = ind_prod_comp.copy()

# Suppression de colonnes non utiles à ce projet

colonnes_to_drop2 = ['Identifiant du message','region_abr','date_heure_publication', 'type_indisponibilite', 'duree_inter_minutes','filiere_production',
                     'type_evenement', 'nom_centrale', 'societe', 'code_eic_societe','cause_arret','remarques', 'statut','unite_de_mesure','code_localisation']
ind_prod_comp.drop(colonnes_to_drop2, axis = 1, inplace =True)

ind_prod_comp.head()

ind_prod_comp.info()

print(ind_prod_comp.columns.tolist())

file_export ='ind_prod_comp200.csv'

ind_prod_comp.head(200).to_csv(file_export)
print("200 premières lignes du dataframe ind_prod_comp exporté avec succès dans le fichier 'ind_prod_comp.csv'.")

# ========================== VENTILATION DES LIGNES DU DATAFRAME PAR DATES FIXES COMPRISES ENTRE '2013-01-01 00:00:00' et  '2022-12-31 23:30:00' PAR PAS DE 00H30

from tqdm import tqdm

# Conversion des colonnes 'date_debut' et 'date_fin_previsionnelle' en datetime
ind_prod_comp['date_debut'] = pd.to_datetime(ind_prod_comp['date_debut'])
ind_prod_comp['date_fin_previsionnelle'] = pd.to_datetime(ind_prod_comp['date_fin_previsionnelle'])

# Création de la liste de dates incrémentées de 30 minutes
liste_dates = pd.date_range(start='2013-01-01 00:00:00', end='2022-12-31 23:30:00', freq='30min')

# Création de la liste pour stocker les nouvelles lignes
new_rows = []

# Parcours de chaque ligne de ind_prod_comp
for idx, row in tqdm(ind_prod_comp.iterrows(), total=ind_prod_comp.shape[0]):
    date_debut = row['date_debut']
    date_fin_previsionnelle = row['date_fin_previsionnelle']

    # Détermination de la borne inférieure et supérieure des dates
    borne_inf_index = liste_dates.searchsorted(date_debut, side='left')
    borne_sup_index = liste_dates.searchsorted(date_fin_previsionnelle, side='right')

    # Vérification que les indices sont dans les limites
    if borne_inf_index >= len(liste_dates):
        borne_inf_index = len(liste_dates) - 1
    if borne_sup_index >= len(liste_dates):
        borne_sup_index = len(liste_dates) - 1

    # Création d'une série de dates comprises entre borne_inf et borne_sup
    dates_range = liste_dates[borne_inf_index:borne_sup_index]

    # Copie de la ligne avec toutes les dates dans new_rows
    new_row = row.drop(['date_debut', 'date_fin_previsionnelle']).to_dict()
    for date in dates_range:
        new_row['date'] = date
        new_rows.append(new_row.copy())

# Création  d'un DataFrame à partir de la liste de nouvelles lignes
df_result_ind_prod = pd.DataFrame(new_rows)

# Sauvegarde du DataFrame résultant dans un fichier CSV
output_path = 'df_result_ind_prod.csv'
df_result_ind_prod.to_csv(output_path, index=False)

output_path


print("FIN CODE VENTILATION")
