==> 4-2-Prefusion_indisponibilites_proc.ipynb <==
 # Préparation du fichier preprocessé des indisponibilités des centrales de productionavant fusion avec le fichier des capacités de production

"""
Created on 2024-06-23

@author: Thierry ALLEM
"""

import numpy as np
import pandas as pd
import seaborn as sns
import datetime
import matplotlib.pyplot as plt

#==========================  CHARGEMENT des fichiers d'indisponibilités des moyens de production de 2013 à 2022 ===============================================================================
df_result_ind_prod = pd.read_csv('df_result_ind_prod.csv', sep=',',encoding='latin-1')
df_result_ind_prod.info()

# Conversion de la colonne 'date' en datetime
df_result_ind_prod['date_debut'] = pd.to_datetime(df_result_ind_prod['date'])
df_result_ind_prod.info()

df_result_ind_prod.isna().sum()

# ===============================RECHERCHE ET TRAITEMENT DES DOUBLONS

doublons = df_result_ind_prod.duplicated(subset=['region', 'filiere_globale', 'code_eic_centrale', 'date']).sum()

print(f"Nombre total de lignes en doublons : {doublons}")

#    Filtration des doublons et copie dans un nouveau DataFrame
nouveau_dataframe_doublons =df_result_ind_prod[df_result_ind_prod.duplicated(subset=['region', 'filiere_globale', 'code_eic_centrale', 'date'], keep=False)]

#   Export des 1000 premières lignes du DataFrame original
file_export = 'df_result_ind_prod1000.csv'
nouveau_dataframe_doublons.head(10000).to_csv(file_export, index=False)
print("1000 premières lignes du dataframe ind_prod_comp exporté avec succès dans le fichier 'ind_prod_comp1000.csv'.")

#   Suppression des doublons en ne conservant que la 1ère occurence
df_result_ind_prod_clean = df_result_ind_prod.drop_duplicates(subset=['region', 'filiere_globale', 'code_eic_centrale', 'date'], keep='first')

# Afficher le nombre de lignes après suppression des doublons
print(f"Nombre de lignes après suppression des doublons : {len(df_result_ind_prod_clean)}")

doublons2 = df_result_ind_prod_clean.duplicated(subset=['region', 'filiere_globale', 'code_eic_centrale', 'date']).sum()

print(f"Nombre total de lignes en doublons après nettoyage: {doublons2}")

df_result_ind_prod_clean.info()

# Regroupement par 'region', 'filiere_globale' et 'date' en sommant les  valeurs de 'capacite_indisponible' et 'capacite_disponible'

df_result_ind_prod_prefusion = df_result_ind_prod_clean.groupby(['region', 'filiere_globale', 'date'])[['capacite_indisponible']].sum().reset_index()

# Renommage de la colonne des filières
dictionnaire= {'filiere_globale':'filiere'}
df_result_ind_prod_prefusion=df_result_ind_prod_prefusion.rename(dictionnaire, axis = 1)

# Le résultat est un nouveau dataframe 'grouped_df' avec les valeurs sommées
print(df_result_ind_prod_prefusion)

#   Export du dataframe néttoyé et regroupé  qui sera ensuite fusionné avec eco2mix et cap_prod
file_export = 'df_result_ind_prod_prefusion.csv'
df_result_ind_prod_prefusion.to_csv(file_export, index=False, sep = ';')
print("Dataframe 'df_result_ind_prod_fusion' exporté avec succès dans le fichier 'df_result_ind_prod_prefusion.csv'.")

# Export des 100 premières lignes dans un nouveau fichier CSV pour aperçu
df_result_ind_prod_prefusion_head = df_result_ind_prod_prefusion.head(100)
df_result_ind_prod_prefusion_head.to_csv('df_result_ind_prod_prefusion_first_100_lines.csv', sep=';', encoding='latin-1', index=False)

print("FIN CODE VENTILATION")
